git-svn-id: svn://192.168.1.10/public/CoffeeMud@10594 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/MiscTech/StdElecPanel.java
@@ -178,7 +178,7 @@ public class StdElecPanel extends StdElecContainer implements Electronics.ElecPa
 							if(powerRemaining>0)
 							{
     							double pctToTake=CMath.div(((((Electronics)I).powerNeeds()<=0)?1:((Electronics)I).powerNeeds()),totalPowerReq);
-    							powerToTake=(int)Math.round(pctToTake * (double)powerRemaining);
+    							powerToTake=(int)Math.round(pctToTake * powerRemaining);
     							if(powerToTake<1)
     								powerToTake=1;
 							}
